opti images + dotenv

+++ controllers/book.js
@@ -30,47 +30,51 @@ exports.modifyBook = (req, res, next) => {
     : { ...req.body };
 
   delete bookObject._userId;
-  Book.findOne({ _id: req.params.id })
+  Book.findById(req.params.id)
+
     .then((book) => {
       if (book.userId != req.auth.userId) {
-        res.status(401).json({ message: "Not authorized" });
-      } else {
-        Book.updateOne(
-          { _id: req.params.id },
-          { ...bookObject, _id: req.params.id }
-        )
-          .then(() => res.status(200).json({ message: "Objet modifié!" }))
-          .catch((error) => res.status(401).json({ error }));
+        return res.status(403).json({ message: "403: unauthorized request" });
       }
+
+      // Empêcher la modification du userId
+      const updatedBook = { ...bookObject, userId: book.userId };
+
+      Book.updateOne({ _id: req.params.id }, updatedBook)
+        .then(() => res.status(200).json({ message: "Objet modifié!" }))
+        .catch((error) => res.status(400).json({ error }));
     })
-    .catch((error) => {
-      res.status(400).json({ error });
-    });
+    .catch((error) => res.status(400).json({ error }));
 };
 
 exports.deleteBook = (req, res, next) => {
-  Book.findOne({ _id: req.params.id })
+  Book.findById(req.params.id)
+
     .then((book) => {
-      if (book.userId != req.auth.userId) {
-        res.status(401).json({ message: "Not authorized" });
-      } else {
-        const filename = book.imageUrl.split("/images/")[1];
-        fs.unlink(`images/${filename}`, () => {
-          Book.deleteOne({ _id: req.params.id })
-            .then(() => {
-              res.status(200).json({ message: "Objet supprimé !" });
-            })
-            .catch((error) => res.status(401).json({ error }));
-        });
+      if (!book) {
+        return res.status(404).json({ message: "Livre non trouvé" });
       }
+      if (book.userId != req.auth.userId) {
+        return res.status(403).json({ message: "403: unauthorized request" });
+      }
+
+      const filename = book.imageUrl.split("/images/")[1];
+      fs.unlink(`images/${filename}`, (err) => {
+        if (err) {
+          console.error("Erreur suppression fichier :", err);
+        }
+
+        Book.deleteOne({ _id: req.params.id })
+          .then(() => res.status(200).json({ message: "Objet supprimé !" }))
+          .catch((error) => res.status(500).json({ error })); // Correction ici
+      });
     })
-    .catch((error) => {
-      res.status(500).json({ error });
-    });
+    .catch((error) => res.status(500).json({ error }));
 };
 
 exports.getOneBook = (req, res, next) => {
-  Book.findOne({ _id: req.params.id })
+  Book.findById(req.params.id)
+
     .then((book) => res.status(200).json(book))
     .catch((error) => res.status(404).json({ error }));
 };
